requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
celery[redis]==5.4.0
redis==5.2.1
pydantic>=2.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
pyjwt==2.10.1
bcrypt==4.2.1
httpx==0.28.1
holehe==1.61
maigret==0.4.4
